Refactor:Remove unused methods
