--- src/main/java/Constants/ErrorMessages.java
@@ -1,4 +1,0 @@
-package Constants;
-
-public class ErrorMessages {
-}
