+++ src/main/java/Constants/Messages.java
@@ -1,6 +1,6 @@
 package Constants;
 
-public class SuccessfulMessages {
+public class Messages {
 
     public static final String EXPECTED_SUCCESSFUL_ADD_TO_CART_MODAL_MESSAGE = "Added!";
     public static final String EXPECTED_SUCCESSFUL_PRODUCT_REVIEW_MESSAGE = "Thank you for your review.";
